CRM# add new component "ModalComponents"
